fix(mobile): make sure logging out clears ephemeral information (current ride, etc.)
